requirements.txt
apipkg==1.5
appdirs==1.4.3
asn1crypto==1.3.0
attrs==19.3.0
bleach==3.1.4
cement==3.0.4
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
colorlog==4.1.0
coverage==5.1
cryptography==2.8
distlib==0.3.0
docutils==0.16
execnet==1.7.1
filelock==3.0.12
idna==2.8
IotaCli==0.0.1
jeepney==0.4.3
Jinja2==2.11.2
keyring==21.2.0
libusb1==1.7.1
MarkupSafe==1.1.1
more-itertools==8.2.0
mysql-connector==2.2.9
ndeflib==0.3.3
nfcpy==1.0.3
packaging==20.3
pathlib==1.0.1
phx-class-registry==3.0.5
phx-filters==2.0.2
pkginfo==1.5.0.1
pluggy==0.13.1
py==1.8.1
pycparser==2.19
pyDes==2.0.1
Pygments==2.6.1
pyOpenSSL==19.1.0
PyOTA==2.1.0
PyOTA-CCurl==1.0.9
PyOTA-PoW==1.1.0
pyparsing==2.4.7
pyserial==3.4
pysha3==1.0.2
pytest==5.4.1
pytest-cov==2.8.1
pytest-forked==1.1.3
pytest-xdist==1.31.0
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3.1
readme-renderer==25.0
regex==2019.11.1
requests==2.22.0
requests-toolbelt==0.9.1
SecretStorage==3.1.2
six==1.14.0
tinydb==3.15.2
toml==0.10.0
tox==3.14.6
tqdm==4.45.0
twine==3.1.1
urllib3==1.25.7
virtualenv==20.0.17
wcwidth==0.1.9
webencodings==0.5.1
